trabajo final

+++ structs.c
@@ -2,7 +2,7 @@ typedef struct pix{
 	float R;
 	float G;
 	float B;
-	float alpha;
+	int alpha;
 	int Y;
 	int binPixel;
 }pixel;
